Split the crate into a library and executable so we can easily test it
